pyproject.toml
[project]
name = "letscrackit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.63.0",
    "google-generativeai>=0.8.6",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "streamlit>=1.54.0",
]
